remove some warnings

+++ kalibri/include/kbclass.h
@@ -733,7 +733,7 @@ protected:
         if (name == 0)
             name = _SC("constructor");
         else alternative_global = true;
-        string overloadName = SqOverloadName::Get(name, nParams);
+        string overloadName = SqOverloadName::Get(name, static_cast<int>(nParams));
 
         if (!alternative_global )
         {
